requirements.txt
alembic==1.13.1
Babel==2.14.0
bcrypt==4.1.2
blinker==1.7.0
cachelib==0.10.2
click==8.0.0
config==0.5.1
dnspython==2.5.0
dominate==2.5.2
email-validator==2.1.0.post1
Flask==2.2.0
Flask-Admin==1.6.1
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
flask-mysql-connector==1.1.0
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-Session==0.6.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.0.0
Jinja2==3.0.0
Mako==1.3.1
MarkupSafe==2.1.1
mysql-connector-python==8.3.0
numpy==1.26.3
pandas==2.2.0
passlib==1.7.4
random-password-generator==2.2.0
passwordgenerator==1.5.1
python-dateutil==2.8.2
pytz==2023.3.post1
random-password-generator==2.2.0
six==1.16.0
speaklater==1.3
SQLAlchemy==1.3.20
typing_extensions==4.9.0
tzdata==2023.4
visitor==0.1.3
Werkzeug==2.2.0
WTForms==2.3.3
zipp==3.17.0
scikit-learn==1.3.0